create separate file for test data for SearchViewModelTest file + delete ExampleUnitTest
