+++ viewmodel/src/test/java/com/cairosquad/viewmodel/searchviewmodel/SearchViewModelTest.kt
@@ -5,9 +5,6 @@ import com.cairosquad.domain.search.usecase.GetExploreMoreUseCase
 import com.cairosquad.domain.search.usecase.GetForYouUseCase
 import com.cairosquad.domain.search.usecase.GetRecentSearchUseCase
 import com.cairosquad.domain.search.usecase.SearchUseCase
-import com.cairosquad.entity.Artist
-import com.cairosquad.entity.Movie
-import com.cairosquad.entity.Series
 import com.google.common.truth.Truth.assertThat
 import io.mockk.coEvery
 import io.mockk.mockk
@@ -51,33 +48,6 @@ class SearchViewModelTest {
             getForYouUseCase
         )
     }
-
-    val movie1 = Movie(
-        id = 1,
-        title = "The dark knight",
-        rating = 4.0f,
-        posterPath = "/img.jpg"
-    )
-
-    val movie2 = Movie(
-        id = 2,
-        title = "Girl",
-        rating = 4.5f,
-        posterPath = "/img.jpg"
-    )
-
-    val series = Series(
-        id = 1,
-        title = "Series",
-        rating = 3.5f,
-        posterPath = "/img.jpg"
-    )
-
-    val artist = Artist(
-        id = 1,
-        name = "Artist",
-        photoPath = "/img.jpg"
-    )
 
     @Test
     fun loadDiscoverMoviesLoadsDataSuccessfully() = runBlocking {
